+++ adminer/lang/cs.inc.php
@@ -303,6 +303,7 @@ Lang::$translations = array(
 	'Lists' => 'Seznamy',
 	'Network' => 'Síť',
 	'Geometry' => 'Geometrie',
+	'Ranges' => 'Rozsahy',
 	'Relations' => 'Vztahy',
 
 	'Editor' => 'Editor',