+++ adminer/lang/xx.inc.php
@@ -303,6 +303,7 @@ Lang::$translations = array(
 	'Lists' => 'Xx',
 	'Network' => 'Xx',
 	'Geometry' => 'Xx',
+	'Ranges' => 'Xx',
 	'Relations' => 'Xx',
 
 	'Editor' => 'Xx',