+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - PostgreSQL: Do not show functions and types created by extensions
 - PostgreSQL: Fix altering functions with parameters of array or user-defined types
 - PostgreSQL: Hide table actions and row editing in pg_catalog and pg_toast
+- PostgreSQL: Support range types (bug #1308)
 - PostgreSQL: Show composite, domain and range types next to enums
 - PostgreSQL: Alter and rename user defined types
 - PostgreSQL: Add values to an enum type without re-creating it